--- chilopod-spider/src/main/java/cn/dtvalley/chilopod/spider/App.java
@@ -1,11 +1,0 @@
-package cn.dtvalley.chilopod.spider;
-
-/**
- * Hello world!
- *
- */
-public class App {
-    public static void main( String[] args ) {
-        System.out.println( "Hello World!" );
-    }
-}
