+++ chilopod-spider/src/main/java/cn/dtvalley/chilopod/spider/Bootstrap.java
@@ -1,0 +1,16 @@
+package cn.dtvalley.chilopod.spider;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * 项目启动入口
+ */
+@SpringBootApplication
+@Slf4j
+public class Bootstrap {
+    public static void main(String[] args) {
+        SpringApplication.run(Bootstrap.class, args);
+    }
+}
