+++ chilopod-spider/src/main/resources/application-dev.yml
@@ -1,0 +1,6 @@
+spring:
+  datasource:
+    url: jdbc:mysql://5628e0ad1c559.sh.cdb.myqcloud.com:9993/b2c?characterEncoding=utf8&useSSL=true
+    username: cdb_outerroot
+    password: shandelzare123
+    driver-class-name: com.mysql.jdbc.Driver
