+++ chilopod-spider/src/main/resources/banner.txt
@@ -1,0 +1,11 @@
+
+       _     _ _                       _
+      | |   (_) |                     | |
+   ___| |__  _| | ___  _ __   ___   __| |
+  / __| '_ \| | |/ _ \| '_ \ / _ \ / _` |
+ | (__| | | | | | (_) | |_) | (_) | (_| |
+  \___|_| |_|_|_|\___/| .__/ \___/ \__,_|
+                      | |
+                      |_|
+
+:: Spring Boot :: ${spring-boot.version}
